+++ CHANGELOG.md
@@ -15,7 +15,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :bug: Bug Fix
 
-* [Fixed a wrong refactoring in `rc48` `super/i-data`](https://github.com/V4Fire/Client/pull/326)
+* [Fixed wrong refactoring in `rc48` `super/i-data`](https://github.com/V4Fire/Client/pull/326)
 
 ## v3.0.0-rc.60 (2020-09-01)