+++ CHANGELOG.md
@@ -17,6 +17,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * [Fixed wrong refactoring in `rc48` `super/i-data`](https://github.com/V4Fire/Client/pull/326)
 
+#### :house: Internal
+
+* Updated dependencies: `@v4fire/core@3.25.1`
+
 ## v3.0.0-rc.60 (2020-09-01)
 
 #### :rocket: New Feature